Conversion from MIR to TIR for type checking / inference

+++ src/frontend/type_system/tir_ast_expressions.rs
@@ -1,6 +1,6 @@
 pub enum TIRExpression {
-    Integer(i128),
-    Float(f64),
+    Integer,
+    Float,
     VariableReference {
         name: String,
     },
